move terminal related state out of ui and into new terminal-window store
